Put Constraint and VariableDefinition into own files
